Add walking movement to camera when keypressed

+++ CustomEngine/src/Engine/Mesh.h
@@ -95,7 +95,7 @@ public:
 	void setBbox(); 
 	SpaceEngine::boundingBox getBbox() const { return bbox; };
 
-	unsigned int getNumTri() { return vertices.size() / 3; }
+	unsigned int getNumTri() { return (int)vertices.size() / 3; }
 
 	// Uniform 
 	glm::mat4 getModelView() const { return model_view; }
